# api/openapi.yaml
openapi: "3.1.0"
info:
  title: korauth — OpenKor Identity / IAM Service
  version: "1.0.0"
  description: |
    JWT (RS256) based authentication and access management.
    All requests come through the `korapi` gateway; this service is not directly exposed.
servers:
  - url: http://localhost:8081
    description: Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "RS256 signed JWT access token. Header - Authorization: Bearer <token>"

  # -------------------------------------------------------------------------
  # Reusable schemas
  # -------------------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              description: Machine-readable, SCREAMING_SNAKE_CASE
              example: USER_NOT_FOUND
            message:
              type: string
              example: "User not found"
            details:
              type: object
              additionalProperties: true
              nullable: true
            request_id:
              type: string
              format: uuid
...
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token:
          type: string
          description: RS256 signed JWT (valid for 15 minutes)
        refresh_token:
          type: string
          description: Opaque random string (valid for 7 days)
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Access token lifetime in seconds
          example: 900
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    UserBase:
      type: object
      required: [username, tenant_id]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
          example: alice
        first_name:
          type: string
          nullable: true
          maxLength: 128
          example: Alice
        last_name:
          type: string
          nullable: true
          maxLength: 128
          example: Smith
        email:
          type: string
          nullable: true
          format: email
          example: alice@example.com
        tenant_id:
          type: string
          format: uuid
          example: "7f3d1c00-0000-0000-0000-000000000001"
    UserResponse:
      allOf:
        - $ref: "#/components/schemas/UserBase"
        - type: object
          required: [id, created_at, updated_at, roles]
          properties:
            id:
              type: string
              format: uuid
            roles:
              type: array
              items:
                $ref: "#/components/schemas/RoleName"
              minItems: 1
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...
    RoleName:
      type: string
      enum: [admin, operator, viewer]
...
    AssignRolesRequest:
      type: object
      required: [roles]
      properties:
        roles:
          type: array
          items:
            $ref: "#/components/schemas/RoleName"
          minItems: 1
          description: |
            Roles to assign to the user. In MVP, single element expected.
            Existing roles are **replaced** (PUT semantics).

  # -------------------------------------------------------------------------
  # Reusable response templates
  # -------------------------------------------------------------------------
  responses:
    Unauthorized:
      description: Token missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: UNAUTHORIZED
              message: "Missing or invalid token"
              request_id: "550e8400-e29b-41d4-a716-446655440000"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: FORBIDDEN
              message: "Insufficient permissions"
              request_id: "550e8400-e29b-41d4-a716-446655440000"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Invalid request body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

# ---------------------------------------------------------------------------
# Paths
# ---------------------------------------------------------------------------
paths:
...
  /v1/auth/refresh:
    post:
      operationId: refreshToken
      tags: [auth]
      summary: Get new access token with refresh token
      description: |
        Token rotation is applied: the used refresh token is invalidated,
        and a new access token is returned along with a new refresh token.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          description: Refresh token invalid or expired
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error:
                  code: INVALID_REFRESH_TOKEN
                  message: "Refresh token is invalid or expired"
                  request_id: "550e8400-e29b-41d4-a716-446655440000"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/auth/users/{id}/roles:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    put:
      operationId: assignRoles
      tags: [roles]
      summary: Assign roles to user (replaces existing roles)
      description: |
        Only `admin` role can access.
        PUT semantics: the provided `roles` array **replaces** all of the user's roles.
        Forbidden to remove own role (403).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignRolesRequest"
      responses:
        "200":
          description: Updated user information (with role list)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

  # =========================================================================
  # Role definitions (seed list)
  # =========================================================================
...